% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoristic.ref.R
\name{aoristic.ref}
\alias{aoristic.ref}
\title{Hour of the week reference chart
Creates and views a simple data frame with reference hour of the week. 
This is a reference for the aoristic.map function that requires a user-selected hour
in order to map spatial distribution for the requested one hour time period.}
\usage{
aoristic.ref()
}
\value{
A data frame with hours of the week referenced
}
\description{
NOTE: The charts from aoristic.summary do not use the same day/hour layout - they move Sunday 
to the end of the layout to make the weekend period clearer to the user.
}
\references{
Ratcliffe, J. H. (2002). Aoristic Signatures and the Spatio-Temporal Analysis of High Volume Crime Patterns. Journal of Quantitative Criminology, 18(1), 23-43.
}
