% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoristic2.df.R
\name{aoristic2.df}
\alias{aoristic2.df}
\title{Create a data.frame with hourly aoristic values}
\usage{
aoristic2.df(data1, Xcoord, Ycoord, DateTimeFrom, DateTimeTo)
}
\arguments{
\item{data1}{data.frame with a minimum of 4 columns with X, Y coords, Start and End date/time}

\item{Xcoord}{a vector of the X coordinate or latitude (passed through for user)}

\item{Ycoord}{a vector of the Y coordinate or longitude (passed through for user)}

\item{DateTimeFrom}{a  vector of the column name for FromDateTime (POSIXct date-time object)}

\item{DateTimeTo}{a vector of the column name for ToDateTime (POSIXct date-time object). If missing, one hour duration assigned.}
}
\value{
A data frame with aoristic values for each hour of the week for each input row
}
\description{
Create a data.frame with hourly aoristic values
}
\references{
Ratcliffe, J. H. (2002). Aoristic Signatures and the Spatio-Temporal Analysis of High Volume Crime Patterns. Journal of Quantitative Criminology, 18(1), 23-43.
}
