% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoristic2.graph.R
\name{aoristic2.graph}
\alias{aoristic2.graph}
\title{Creates eight charts for each day of the week (and a total) based on aoristic distribution}
\usage{
aoristic2.graph(data1, marks = "")
}
\arguments{
\item{data1}{data.frame with the output from aoristic2.df}

\item{marks}{marks='T' to show tick marks for week distribution. Leave blank otherwise.}
}
\description{
Creates eight charts for each day of the week (and a total) based on aoristic distribution
}
\references{
Ratcliffe, J. H. (2002). Aoristic Signatures and the Spatio-Temporal Analysis of High Volume Crime Patterns. Journal of Quantitative Criminology, 18(1), 23-43.
}
