% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoristic2.map.R
\name{aoristic2.map}
\alias{aoristic2.map}
\title{Creates a plot showing aoristic values for one hour of the week}
\usage{
aoristic2.map(data1, AorHour = "1")
}
\arguments{
\item{data1}{data.frame created by aoristic2.df}

\item{AorHour}{use seelected number for an hour in the week (range 1-168)}
}
\value{
A ggplot object
}
\description{
Creates a plot showing aoristic values for one hour of the week
}
\references{
Ratcliffe, J. H. (2002). Aoristic Signatures and the Spatio-Temporal Analysis of High Volume Crime Patterns. Journal of Quantitative Criminology, 18(1), 23-43.
}
