% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CouncilDistrict}
\alias{CouncilDistrict}
\title{Council district (spatial polygon data frame)}
\source{
City of Arlington
}
\usage{
CouncilDistrict
}
\description{
A spatial polygon data frame of council districts in the Arlington PD jurisdiction.
}
\author{
George Kikuchi, 2013-09-13
}

