% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dcburgsum}
\alias{dcburgsum}
\title{Summary output from aoristic analysis of DC burglaries}
\format{
A data frame with 24 observations and 8 variables:
\describe{
  \item{Range}{The hour represented by the column, in 24 hour format}
  \item{Mon}{Sum of aoristic values for Mondays by hour}
  \item{Tue}{Sum of aoristic values for Tuesdays by hour}
  \item{Wed}{Sum of aoristic values for Wednesdays by hour}
  \item{Thu}{Sum of aoristic values for Thursdays by hour}
  \item{Fri}{Sum of aoristic values for Fridays by hour}
  \item{Sat}{Sum of aoristic values for Saturdays by hour}
  \item{Sun}{Sum of aoristic values for Sundays by hour}
}
}
\usage{
dcburgsum
}
\description{
A dataset containing the summarized output generated by aoristic.summary() 
on the data frame produced by the aoristic function aoristic.df(dcburglaries, ...).
}
\keyword{datasets}
