% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyrobbery}
\alias{nyrobbery}
\title{Robberies, Manhattan, (New York), 2019}
\format{
A data frame with 3089 rows and 10 variables:
\describe{
  \item{CMPLNT_FR_DT}{Exact date of occurrence for the reported event (or starting date of occurrence, if CMPLNT_TO_DT exists)}
  \item{CMPLNT_FR_TM}{Exact time of occurrence for the reported event (or starting time of occurrence, if CMPLNT_TO_TM exists)}
  \item{CMPLNT_TO_DT}{Ending date of occurrence for the reported event, if exact time of occurrence is unknown}
  \item{CMPLNT_TO_TM}{Ending time of occurrence for the reported event, if exact time of occurrence is unknown}
  \item{X_COORD_CD}{X-coordinate for New York State Plane Coordinate System, Long Island Zone, NAD 83, units feet (FIPS 3104)}
  \item{Y_COORD_CD}{Y-coordinate for New York State Plane Coordinate System, Long Island Zone, NAD 83, units feet (FIPS 3104)}
  \item{Latitude}{Midblock Latitude coordinate for Global Coordinate System, WGS 1984, decimal degrees (EPSG 4326)}
  \item{Longitude}{Midblock Longitude coordinate for Global Coordinate System, WGS 1984, decimal degrees (EPSG 4326)}
  \item{STARTDateTime}{Start date and time of the burglary in POSIXct format}
  \item{ENDDateTime}{End date and time of the burglary in POSIXct format}
}
}
\source{
\url{https://data.cityofnewyork.us/Public-Safety/NYPD-Complaint-Data-Current-Year-To-Date-/5uac-w243/data}
}
\usage{
nyrobbery
}
\description{
A dataset containing felony robberies in Manhattan Borough, New York City with a start
or event date in 2019. Also contains coordinates in X, Y format using New York State Plane
Coordinate System, Long Island Zone, NAD 83, units feet (FIPS 3104). The data also include 
midblock Latitude and longitude coordinates in WGS 1984, decimal degrees (EPSG 4326). 
Original source data contained 35 fields. This is a filtered dataset.
}
\keyword{datasets}
