\docType{package}
\name{aoristic-package}
\alias{aoristic-package}
\title{Creating a kml file with aoristic graph: Aoristic Analysis with spatial kml data}
\description{
Using crime incident data with lat/lon, DateTimeFrom, and
DateTimeTo, a total of three (3) aoristic graphs can be
created: 1) density and contour; 2) grid count; and 3)
shapefile boundary.
}
\details{
Google Earth needs to be installed in order to view a kml
output with aoristic graphs. aoristic.grid produces a kml
file of aoristic graphs based on a 5*5 grid. aoristic.shp
produces a kml file of aoristic graphs based on input
shapefile boundary aoristic.density produces a kml file of
aoristic graphs with kernel density
}
\author{
George Kikuchi \email{gkikuchi@csufresno.edu}
}
\references{
Ratcliffe, J. H. (2002). Aoristic Signatures and the
Spatio-Temporal Analysis of High Volume Crime Patterns.
Journal of Quantitative Criminology, 18(1), 23-43.
}
\keyword{aoristic}

