% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{aoristic.all.graph}
\alias{aoristic.all.graph}
\title{Creating a kml file with aoristic graph}
\usage{
aoristic.all.graph(...)
}
\description{
This function is defunct
}
\keyword{Creating}
\keyword{This}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{aoristic}
\keyword{data}
\keyword{defunct}
\keyword{for}
\keyword{frame}
\keyword{function}
\keyword{graph}
\keyword{internal}
\keyword{is}
\keyword{using}
