% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoristic.datacheck.R
\name{aoristic.datacheck}
\alias{aoristic.datacheck}
\title{Check input data}
\usage{
aoristic.datacheck(data1, Xcoord, Ycoord, DateTimeFrom, DateTimeTo)
}
\arguments{
\item{data1}{data.frame with a minimum of 4 columns with X, Y coords, Start and End date/time}

\item{Xcoord}{a vector of the X coordinate or latitude (passed through for user)}

\item{Ycoord}{a vector of the Y coordinate or longitude (passed through for user)}

\item{DateTimeFrom}{a  vector of the column name for FromDateTime (POSIXct date-time object)}

\item{DateTimeTo}{a vector of the column name for ToDateTime (POSIXct date-time object). If missing, one hour duration assigned.}
}
\value{
A data frame flagging any problems or logical errors from an aoristic data check
}
\description{
The aoristic.df function handles data with Start (or From) and End (or To) datetime objects. Sometimes
these data can be messy or inaccurately recorded (such as crime data from some police departments). This
function checks for two common problems and reports the result in a column labeled 'aoristic_datacheck'.
Rows with missing data for the End (or To) event are flagged with '1'.
}
\details{
Rows with illogical data where the #' end datetime occurs before the start datetime are 
flagged '2'. 

Note that the aoristic.df function will #' still function with these 
problems, but events flagged '1' will be assigned a default time span of one
hour, and events flagged '2' will be skipped.
}
\examples{
datacheck.df <- aoristic.datacheck(dcburglaries, "X", "Y", "StartDateTime", "EndDateTime")
}
\references{
Ratcliffe, J. H. (2002). Aoristic Signatures and the Spatio-Temporal Analysis of High Volume Crime Patterns. Journal of Quantitative Criminology, 18(1), 23-43.
}
