\docType{data}
\name{aoristic}
\alias{aoristic}
\alias{aoristicData}
\title{crime (df) and council district data (spdf)}
\source{
  City of Arlington
}
\description{
  A burglary incident data and spatial polygon data frame
  of council districts.
}
\author{
  George Kikuchi, 2013-09-13
}
\keyword{datasets}

