\docType{data}
\name{aoristic}
\alias{aoristic}
<<<<<<< HEAD
\alias{aoristicSampleData}
\title{Sample data of crime (df) and council district (spdf)}
=======
\alias{aoristicData}
\title{crime (df) and council district data (spdf)}
>>>>>>> 1eb2df43f44e523adc07872cb53ad8132adea5b3
\source{
  City of Arlington
}
\description{
  Sample data of burglary incident data frame and spatial
  polygon data frame of council districts.
}
\author{
  George Kikuchi, 2013-09-13
}
\keyword{datasets}

