% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoristic2.ref.R
\name{aoristic2.ref}
\alias{aoristic2.ref}
\title{Creates a data.frame with reference hour of the week (ref purposes only)}
\usage{
aoristic2.ref()
}
\value{
A data frame with hours of the week referenced
}
\description{
Creates a data.frame with reference hour of the week (ref purposes only)
}
\references{
Ratcliffe, J. H. (2002). Aoristic Signatures and the Spatio-Temporal Analysis of High Volume Crime Patterns. Journal of Quantitative Criminology, 18(1), 23-43.
}
