% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoristic2.summary.R
\name{aoristic2.summary}
\alias{aoristic2.summary}
\title{Summarizes the aoristic sums for each hour of the week, based on aoristic data frame}
\usage{
aoristic2.summary(data1, output = "")
}
\arguments{
\item{data1}{data.frame with output from aoristic function}

\item{output}{enter 'xlsx' for an Excel format output, 'jpg' for JPEG grid, blank otherwise}
}
\value{
A data frame with aoristic values summed for each hour of the week
}
\description{
Summarizes the aoristic sums for each hour of the week, based on aoristic data frame
}
\references{
Ratcliffe, J. H. (2002). Aoristic Signatures and the Spatio-Temporal Analysis of High Volume Crime Patterns. Journal of Quantitative Criminology, 18(1), 23-43.
}
