\docType{data}
\name{arlington}
\alias{arlington}
\title{arlington burglary incident data (data frame)}
\source{
Arlington PD
}
\usage{
arlington
}
\description{
A data frame of burglary incidents in the Arlington PD with
four fields (DateTimeFrom, DateTimeTo, lon/lat).
}
\author{
George Kikuchi, 2013-09-13
}

