% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aoristic-package.R
\docType{data}
\name{CouncilDistrict}
\alias{CouncilDistrict}
\title{Council district (spatial polygon data frame)}
\source{
City of Arlington
}
\usage{
data(aoristic)
}
\description{
A spatial polygon data frame of council districts in the Arlington PD jurisdiction.
}
\author{
George Kikuchi, 2013-09-13
}

