\docType{data}
\name{CouncilDistrict}
\alias{CouncilDistrict}
\title{Council district (spatial polygon data frame)}
\source{
  City of Arlington
}
\description{
  A spatial polygon data frame of council districts in the
  Arlington PD jurisdiction.
}
\author{
  George Kikuchi, 2013-09-13
}
\keyword{datasets}

